Add confirms to Deletes and Text
- Add confirmation messages to Item and box deletion, and to the message
button
- Change user-facing references to 'write access' into 'Admin', as it
better describes that level of resource control (closes Issue#4)

+++ app/routes/boxes/box.js
@@ -8,13 +8,27 @@ export default Route.extend({
   actions: {
     deleteBox (box) {
       // console.log('delete', box.get('name'))
-      box.destroyRecord()
-        .then((response) => {
-          this.toast.success('Box Successfully Removed', 'Success')
-          return response
-        })
-        .then(() => this.transitionTo('boxes'))
-        .catch(() => this.toast.error('Error Removing the Box', 'Failure'))
+      if (confirm('Are you sure you want to delete this box? The box and all of it\'s contents will be destroyed.')) {
+        box.destroyRecord()
+          .then((response) => {
+            this.toast.success('Box Successfully Removed', 'Success')
+            return response
+          })
+          .then(() => this.transitionTo('boxes'))
+          .catch(() => this.toast.error('Error Removing the Box', 'Failure'))
+      }
+    },
+    removeBoxFromReader (box) {
+      // console.log('delete', box.get('name'))
+      if (confirm('Are you sure you want to remove this box from your list?')) {
+        box.destroyRecord()
+          .then((response) => {
+            this.toast.success('Box Successfully Removed', 'Success')
+            return response
+          })
+          .then(() => this.transitionTo('boxes'))
+          .catch(() => this.toast.error('Error Removing the Box', 'Failure'))
+      }
     },
     editBoxView (box) {
       this.transitionTo('/boxes/' + box.get('id') + '/edit')
