+++ app/routes/items/item.js
@@ -15,30 +15,34 @@ export default Route.extend({
     deleteItem (item) {
       // console.log('delete', item.get('name'))
       const box = item.get('box')
-      item.destroyRecord()
-        .then((response) => {
-          this.toast.success('Item Deleted', 'Success')
-          return response
-        })
-        .then(() => this.transitionTo('boxes.box', box))
-        .catch(() => this.toast.error('Error Deleting this Item', 'Failure'))
+      if (confirm('Are you sure you want to permanently delete this item?')) {
+        item.destroyRecord()
+          .then((response) => {
+            this.toast.success('Item Deleted', 'Success')
+            return response
+          })
+          .then(() => this.transitionTo('boxes.box', box))
+          .catch(() => this.toast.error('Error Deleting this Item', 'Failure'))
+        }
     },
     sendMessage (item, permission) {
       const body = `Hi ${permission.get('user.email')}, ${this.get('auth.credentials.email')} sent a reminder about Item: '${item.get('name')}' in Box: '${item.get('box.name')}'!`
-      const data = {
-        twilio: {
-          user_id: permission.get('user.id'),
-          box_id: item.get('box.id'),
-          item_id: item.get('id'),
-          body: 'Hi hello'
+      if (confirm(`Send text: '${body}'?`)) {
+        const data = {
+          twilio: {
+            user_id: permission.get('user.id'),
+            box_id: item.get('box.id'),
+            item_id: item.get('id'),
+            body: 'Hi hello'
+          }
         }
+        this.get('auth').twilioMessage(data)
+          .then((response) => {
+            this.toast.success(body, 'Text Sent! Your message:')
+            return response
+          })
+          .catch(() => this.toast.error(`${permission.get('user.email')} may not have a valid phone number listed.`, 'Text not sent'))
       }
-      this.get('auth').twilioMessage(data)
-        .then((response) => {
-          this.toast.success(body, 'Text Sent! Your message:')
-          return response
-        })
-        .catch(() => this.toast.error(`${permission.get('user.email')} may not have a valid phone number listed.`, 'Text not sent'))
     }
   }
 })
